fix sources card spacing
