+++ src/widgets/copilot/components/Messages/incoming.scss
@@ -33,7 +33,7 @@
   cursor: pointer;
   min-width: 160px;
   max-width: 160px;
-  height: 128px;
+  height: 74px;
   padding: 8px;
   border: 1px solid transparent;
 }
